Rename Transnational to Network Merchants
